sistema/requirements.txt
streamlit
pandas
psycopg2-binary
requests
fpdf2
qrcode
pillow